118 authorization

+++ backend/routes/posts.js
@@ -40,8 +40,11 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
+    // console.log(req.userData);
+    // return res.status(200).json({});
     post.save().then(createdPost => {
       res.status(201).json({
         message: "Post added successfully",
@@ -68,11 +71,20 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      // creator: req.userData.userId
     });
     console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+
+    // FILTER BY DB id and userId !!
+    // mongo check in the request if nModified = 0 OR 1  .    IF 0 then no update done
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      // console.log(result);
+      if( result.nModified > 0 ){
+        res.status(200).json({ message: "Update successful!" });
+      }else{
+        res.status(401).json({ message: "Update not Authorized !" });
+      }
     });
   }
 );
@@ -109,10 +121,15 @@ router.get("/:id", (req, res, next) => {
   });
 });
 
+//MONGO FILTER BY n
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+  Post.deleteOne({ _id: req.params.id , creator: req.userData.userId }).then(result => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+    if( result.n > 0 ){
+      res.status(200).json({ message: " Post deleted successful!" });
+    }else{
+      res.status(401).json({ message: "Post delete not Authorized !" });
+    }
   });
 });
